Move `projects` to `features` directory
