+++ web-admin/src/features/projects/ProjectHero.svelte
