+++ web-admin/src/routes/-/github/connect/request/+page.svelte
@@ -2,14 +2,14 @@
   import { goto } from "$app/navigation";
   import { createAdminServiceGetCurrentUser } from "@rilldata/web-admin/client";
   import { ADMIN_URL } from "@rilldata/web-admin/client/http-client";
-  import Github from "@rilldata/web-common/components/icons/Github.svelte";
   import CodeBlockInline from "@rilldata/web-common/components/calls-to-action/CodeBlockInline.svelte";
   import CtaContentContainer from "@rilldata/web-common/components/calls-to-action/CTAContentContainer.svelte";
   import CtaHeader from "@rilldata/web-common/components/calls-to-action/CTAHeader.svelte";
   import CtaLayoutContainer from "@rilldata/web-common/components/calls-to-action/CTALayoutContainer.svelte";
   import CtaMessage from "@rilldata/web-common/components/calls-to-action/CTAMessage.svelte";
   import KeyboardKey from "@rilldata/web-common/components/calls-to-action/KeyboardKey.svelte";
-  import GithubRepoInline from "../../../../../components/projects/GithubRepoInline.svelte";
+  import Github from "@rilldata/web-common/components/icons/Github.svelte";
+  import GithubRepoInline from "../../../../../features/projects/GithubRepoInline.svelte";
 
   const remote = new URLSearchParams(window.location.search).get("remote");
   const user = createAdminServiceGetCurrentUser({
