+++ web-admin/src/routes/-/github/connect/retry-install/+page.svelte
@@ -12,7 +12,7 @@
   import CtaMessage from "@rilldata/web-common/components/calls-to-action/CTAMessage.svelte";
   import KeyboardKey from "@rilldata/web-common/components/calls-to-action/KeyboardKey.svelte";
   import GithubFail from "@rilldata/web-common/components/icons/GithubFail.svelte";
-  import GithubRepoInline from "../../../../../components/projects/GithubRepoInline.svelte";
+  import GithubRepoInline from "../../../../../features/projects/GithubRepoInline.svelte";
 
   const remote = new URLSearchParams(window.location.search).get("remote");
   const user = createAdminServiceGetCurrentUser({
