+++ web-admin/src/routes/[organization]/[project]/-/logs/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { page } from "$app/stores";
-  import ProjectDeploymentLogs from "@rilldata/web-admin/components/projects/ProjectDeploymentLogs.svelte";
+  import ProjectDeploymentLogs from "@rilldata/web-admin/features/projects/ProjectDeploymentLogs.svelte";
   import VerticalScrollContainer from "@rilldata/web-common/layout/VerticalScrollContainer.svelte";
 
   $: organization = $page.params.organization;
